Add Code of Conduct.
This code is based on Idris Community Standards and similarly to them
suggest to report violation to "any member of coq-community (at your discretion)".
This is far from ideal if we really want people to feel like their violation
reports are welcome so when we are active enough, we should rather do like in
the original version (Julia Community Standards) and have a team of stewards
and a special e-mail address. See: https://julialang.org/community/stewards/